+++ flowmachine/flowmachine/__init__.py
@@ -21,7 +21,7 @@ and inherit from FlowMachine's main `Query()` class.
 from ._version import get_versions
 
 __version__ = get_versions()["version"]
-__flowdb_version__ = "*"
+__flowdb_version__ = "0.2.0"
 
 del get_versions